Improvements 2

+++ src/resources/file/remove.js
@@ -1,17 +1,24 @@
 import {File, FileData} from '../../models';
 import {applyCheckPermissions} from '../../utils-data';
+import isArray from 'lodash/isArray';
 
 function remove({id} = {}) {
     if (!id) {
         return Promise.reject(400);
     }
 
+    id = isArray(id) ? id : [id];
+
+    if(id.length === 0) {
+        return Promise.reject(400);
+    }
+
     return Promise.all([
         File.remove({
-            _id: id
+            _id: {$in: id}
         }).exec(),
         FileData.remove({
-            files_id: id
+            files_id: {$in: id}
         }).exec()
     ]);
 }
